requirements.txt
Django==1.8.4
Pillow==2.9.0
djangorestframework==3.2.3
eventlet==0.17.4
greenlet==0.4.9
gunicorn==19.3.0
psycopg2==2.6.1
python-memcached==1.57
six==1.9.0
